REORG: Splitting classes into multiple ones
vDF -> multiple sub classes
