+++ verticapy/core/vdataframe/eval.py
@@ -1,0 +1,136 @@
+"""
+(c)  Copyright  [2018-2023]  OpenText  or one of its
+affiliates.  Licensed  under  the   Apache  License,
+Version 2.0 (the  "License"); You  may  not use this
+file except in compliance with the License.
+
+You may obtain a copy of the License at:
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless  required  by applicable  law or  agreed to in
+writing, software  distributed  under the  License is
+distributed on an  "AS IS" BASIS,  WITHOUT WARRANTIES
+OR CONDITIONS OF ANY KIND, either express or implied.
+See the  License for the specific  language governing
+permissions and limitations under the License.
+"""
+# Standard Python Modules
+import re
+from typing import Union
+
+# VerticaPy Modules
+from verticapy._utils._collect import save_verticapy_logs
+from verticapy.errors import QueryError
+from verticapy.sql.flex import isvmap
+from verticapy._utils._cast import to_category
+from verticapy.core.str_sql import str_sql
+from verticapy.sql._utils._format import quote_ident
+from verticapy.sql.dtypes import get_data_types
+
+
+class vDFEVAL:
+    def __setattr__(self, attr, val):
+        from verticapy.core.vcolumn import vColumn
+
+        if isinstance(val, (str, str_sql, int, float)) and not isinstance(val, vColumn):
+            val = str(val)
+            if self.is_colname_in(attr):
+                self[attr].apply(func=val)
+            else:
+                self.eval(name=attr, expr=val)
+        elif isinstance(val, vColumn) and not (val.init):
+            final_trans, n = val.init_transf, len(val.transformations)
+            for i in range(1, n):
+                final_trans = val.transformations[i][0].replace("{}", final_trans)
+            self.eval(name=attr, expr=final_trans)
+        else:
+            self.__dict__[attr] = val
+
+    def __setitem__(self, index, val):
+        setattr(self, index, val)
+
+    @save_verticapy_logs
+    def eval(self, name: str, expr: Union[str, str_sql]):
+        """
+    Evaluates a customized expression.
+
+    Parameters
+    ----------
+    name: str
+        Name of the new vColumn.
+    expr: str
+        Expression in pure SQL to use to compute the new feature.
+        For example, 'CASE WHEN "column" > 3 THEN 2 ELSE NULL END' and
+        'POWER("column", 2)' will work.
+
+    Returns
+    -------
+    vDataFrame
+        self
+
+    See Also
+    --------
+    vDataFrame.analytic : Adds a new vColumn to the vDataFrame by using an advanced 
+        analytical function on a specific vColumn.
+        """
+        from verticapy.core.vcolumn import vColumn
+
+        if isinstance(expr, str_sql):
+            expr = str(expr)
+        name = quote_ident(name.replace('"', "_"))
+        if self.is_colname_in(name):
+            raise NameError(
+                f"A vColumn has already the alias {name}.\n"
+                "By changing the parameter 'name', you'll "
+                "be able to solve this issue."
+            )
+        try:
+            query = f"SELECT {expr} AS {name} FROM {self.__genSQL__()} LIMIT 0"
+            ctype = get_data_types(query, name[1:-1].replace("'", "''"),)
+        except:
+            raise QueryError(
+                f"The expression '{expr}' seems to be incorrect.\nBy "
+                "turning on the SQL with the 'set_option' function, "
+                "you'll print the SQL code generation and probably "
+                "see why the evaluation didn't work."
+            )
+        if not (ctype):
+            ctype = "undefined"
+        elif (ctype.lower()[0:12] in ("long varbina", "long varchar")) and (
+            self._VERTICAPY_VARIABLES_["isflex"]
+            or isvmap(expr=f"({query}) VERTICAPY_SUBTABLE", column=name,)
+        ):
+            category = "vmap"
+            ctype = "VMAP(" + "(".join(ctype.split("(")[1:]) if "(" in ctype else "VMAP"
+        else:
+            category = to_category(ctype=ctype)
+        all_cols, max_floor = self.get_columns(), 0
+        for column in all_cols:
+            column_str = column.replace('"', "")
+            if (quote_ident(column) in expr) or (
+                re.search(re.compile(f"\\b{column_str}\\b"), expr)
+            ):
+                max_floor = max(len(self[column].transformations), max_floor)
+        transformations = [
+            (
+                "___VERTICAPY_UNDEFINED___",
+                "___VERTICAPY_UNDEFINED___",
+                "___VERTICAPY_UNDEFINED___",
+            )
+            for i in range(max_floor)
+        ] + [(expr, ctype, category)]
+        new_vColumn = vColumn(name, parent=self, transformations=transformations)
+        setattr(self, name, new_vColumn)
+        setattr(self, name.replace('"', ""), new_vColumn)
+        new_vColumn.init = False
+        new_vColumn.init_transf = name
+        self._VERTICAPY_VARIABLES_["columns"] += [name]
+        self.__add_to_history__(
+            f"[Eval]: A new vColumn {name} was added to the vDataFrame."
+        )
+        return self
+
+
+class vDCEVAL:
+    def __setattr__(self, attr, val):
+        self.__dict__[attr] = val
